Fix babel not matching source files
When I moved the the new src files from `src/js` to `src/scripts` I forgot to make the loader check for that and didn't notice because most things being transpiled already have native browser support.

The loader will not run on files from both folders, and I've updated tests to ensure that the transpilation is actually happening. The simplest way of doing this is by ensuring there are no errors in the output and using some syntax that is not natively supported by webpack. This test will likely need to be updated in the future when webpack's parser `acorn` gains support for the syntax I used.

This test also makes sure that this syntax that should be transpiled does not make it through in that event. It will likely be a few years before browser support is good enough to ship the Object spread operator.